backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
bcrypt==4.1.1
qiskit>=0.43.0
qiskit-aer>=0.13.0
numpy>=1.23.0
pytest>=7.0.0
pytest-cov>=4.0.0
